Implemented triangle based sensor FOV detection.
Because swing uses a not so traditional coordinate system with an upper
left corner for the origin I needed to deal with this. Vector math based
solutions would be much nicer but this distance + triangle based
solution works fine. At the moment the debug info is still enabled.
Still need to implement a few things but the core is ready.